Break out loungeMute backend calls to a new ModerationService
